requirements.txt
pandas>=2.0.0
numpy>=1.24.0
geopandas>=0.14.0
shapely>=2.0.0
pyproj>=3.6.0
rasterio>=1.3.9
folium>=0.15.0
langchain>=0.2.0
chromadb>=0.5.0
sentence-transformers>=2.2.2
transformers>=4.40.0
faiss-cpu>=1.8.0.post1
pillow>=10.0.0
opencv-python>=4.8.0
scikit-image>=0.22.0
matplotlib>=3.8.0
streamlit>=1.35.0
gradio>=4.0.0
streamlit-folium>=0.17.4
torch>=2.1.0
tqdm>=4.66.0
pyyaml>=6.0.1
python-json-logger>=2.0.7
